Reapply changes to _functions
